feat(rad): apply 0.2 tally pre-condition before aggregation
So that try_data_request and the witnet_toolkit behave the same as
an actual node (RadManager)
