+++ node/src/actors/rad_manager/handlers.rs
@@ -1,17 +1,22 @@
 //! Message handlers for `RadManager`
 
-use actix::{Handler, ResponseFuture};
 use std::time::Duration;
+
+use actix::{Handler, ResponseFuture};
+use futures::FutureExt;
+
 use witnet_data_structures::radon_report::{RadonReport, ReportContext};
-use witnet_rad::{error::RadError, script::RadonScriptExecutionSettings, types::RadonTypes};
-use witnet_validations::validations::{
-    evaluate_tally_precondition_clause, run_tally, TallyPreconditionClauseResult,
+use witnet_rad::{
+    conditions::{evaluate_tally_precondition_clause, TallyPreconditionClauseResult},
+    error::RadError,
+    script::RadonScriptExecutionSettings,
+    types::RadonTypes,
 };
+use witnet_validations::validations::run_tally;
 
 use crate::actors::messages::{ResolveRA, RunTally};
 
 use super::RadManager;
-use futures::FutureExt;
 
 // This constant is used to ensure that a RetrievalTimeoutError is committed after 10 seconds
 // This value must be lower than half an epoch, and having enough time to broadcasting the commit.
@@ -118,9 +123,11 @@ impl Handler<RunTally> for RadManager {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::utils::test_actix_system;
     use actix::{Actor, MailboxError, Message};
+
+    use crate::utils::test_actix_system;
+
+    use super::*;
 
     #[test]
     fn rad_manager_handler_can_panic() {
